A fix for https://github.com/svenvc/zinc/issues/118
When the content-length and the actual data available differ (i.e. there are not enough bytes left to read before EOF), which is an error in any case, the response was different between a ZnStringEntity and a ZnByteArrayEntity. Now they behave the same.

Furthermore, in ZnUtils class>>#streamFrom:to:size: there was a read that did not check the amount read.